pyproject.toml
[tool.poetry]
name = "flask-test"
version = "0.1.0"
description = ""
authors = ["guibos <22233599+guibos@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "flask_test"}]

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.8.0"
pydantic = "^2.4.2"
flask = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[tool.pytest.ini_options]
pythonpath = ["src"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
